backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
openai
faiss-cpu
pydantic
python-dotenv
langchain-community